Pet 282 fix : 패밀리 가입 기간 로직 수정

+++ Domain-Module/Todo-Module/Todo-Application/src/main/java/com/pawith/todoapplication/dto/response/TodoTeamInfoResponse.java
@@ -12,5 +12,5 @@ public class TodoTeamInfoResponse {
     private String teamProfileImageUrl;
     private String teamName;
     private Authority authority;
-    private Integer registerPeriod;
+    private Long registerPeriod;
 }
